Dpkg::Build::Info: Avoid whitelist in function name
While the terms whitelist and blacklist do not have an etymology
with racial connotations, they still attribute positive and negative
connotations to these colors for terms that are not explicitly
descriptive of what they mean.

We can use instead terms that are more descriptive, and have no negative
connotations.

Deprecate get_build_env_whitelist() in favor of get_build_env_allowed(),
and emit warnings when using the former.