Dpkg::Shlibs::SymbolFile: Avoid blacklist and whitelist terms
While the terms whitelist and blacklist do not have an etymology
with racial connotations, they still attribute positive and negative
connotations to these colors for terms that are not explicitly
descriptive of what they mean.

We can use instead terms that are more descriptive, and have no negative
connotations.

Explicitly spell out that this is about internal symbols, and whether we
allow them to appear in symbol files. Even though there do not appear to
be users for the old directives in Debian, we will still do an ordered
transition with deprecation warnings in case private and/or third-party
users are affected.