Dpkg::Vendor::Debian: Generalize feature areas defaults
Split setting the build features to use, from acting on them to set the
build flags based on these.

This makes it possible to override these feature defaults by vendors
inheriting from a parent, so that they do not have to mangle the
generated build flags, or need to modify the parent vendor module to be
able to change those defaults.