Add Protected field support
This field is intended to make it possible to move several of the
current packages marked as Essential, so that they can be removed on
installations where these do not make sense being installed.

Protected packages have some of the properties of Essential, but not
all. These are intended to be used mostly for packages that are involved
in booting the system.