Dpkg::Build::Info: Clarify by giving context to the deprecation warning
These built-in warnings provide no context, and in this particular case
the warning message was a bit cryptic anyway. So we need to give it
more context. In this case just fully qualifying the function name, and
stating what is being deprecated, should be enough to give a hint of
what is going on.

Closes: #963844
Reported-by: Sven Joachim <svenjoac@gmx.de>