data: Prefix the specs file spec string self_spec with + instead of *
Using * as a prefix for the spec entry makes it override any previous
setting. This is problematic when we pass two or more -specs options
to the compiler (f.ex. to link and compile on the same run), as then
only the last one will take effect, breaking the builds.

Closes: #870383
Ref: https://dev.gnupg.org/T4982
Prompted-by: NIIBE Yutaka <gniibe@fsij.org>
Tested-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>