Dpkg::Source::Package: Explicitly initialize constructor options
Some of the options had not been explicitly initialized to their
implicit default value, which broke when the code to verify signatures
got refactored into Dpkg::OpenPGP::verify_signature() with a default
for require_valid_signature of 1.

We explicitly initialize all supported options so that this does not
happen again in the future.

Closes: #964017