build: Detect GNU program variants for make, patch and tar in CPAN
On systems where the defaults for these programs are not the GNU
variants, we need to look for them and use them in preference. Search
in the CPAN distribution build system, for the names prefixed with «g»
which is the convention used on BSD systems for GNU variants, and use
them to replace the variables in the Dpkg module.