Dpkg::Source::Package::V1: Explicitly initialize constructor options
Some of the options had not been explicitly initialized to their
implicit default value, which broke when the code to verify signatures
got refactored into Dpkg::OpenPGP::verify_signature() with a default
for require_valid_signature of 1.

This is required here too as this class does not call the initialization
from its parent. This is error prone, and will be refactored in the
future to make use of a common setup in the parent.