Dpkg::Exit: Fix exit handler on program termination
We cannot hook into the __DIE__ pseudo signal handler as then any
eval usage becomes unsafe. Instead we hook into an END code block,
and fix the run_exit_handler code to deplete the @handlers array so
that we do not execute the handlers repeatedly and to reset the
installed handlers as not needed anymore.

Closes: #966083