test: Fix test_command_exec program invocation
From exec(3):

	The argument arg0 should point to a filename string that is
	associated with the process being started by one of the exec
	functions.

Unfortunately, this test sets arg0 to the string "arg 0" this causes the
busybox multicall binary on Alpine Linux to assume that the applet "arg
0" (instead of true) should be executed. However, as such an applet does
not exist, the tests fails. This commit fixes the failing test by
setting arg0 correctly (as other parts of the dpkg codebase using the
command API do too).

Signed-off-by: Guillem Jover <guillem@debian.org>