build: Pre-process the curses header before parsing it
We need to extract the KEY_* macros from the curses header file that
we are going to end up including in the source. But we should not
assume any kind of structure, because that's just fragile, as the
contents can be included from other header files, or be protected
behind a pre-processor macro or similar.

Instead we use the CPP -dD option which will make the pre-processor
handle all the internal details and output a list of active macros
with their values, which we can then safely parse.

Closes: #970545