libdpkg: Fix allocation size wrap around when growing a varbuf
When we allocate the buffer when growing a varbuf, we need to make sure
the new size is not wrapping around, or we might end up allowing to
write into unallocated memory or memory allocated for another buffer.

If we have a buffer that takes half the possibly allocatable memory,
and need less than half of that, we do not make it possible to allocate
just that instead of its double, as that would certainly leave the rest
of the program or the entire system without much resources anyway.

Stable-Candidate: 1.19.x