Support remove-on-upgrade conffile flag via DEBIAN/conffiles
Add support for a new "remove-on-upgrade" flag in DEBIAN/conffiles to
mark that the package wants to remove a given conffile (without having
to rely on maintscripts).

The flag is added to the conffile line before the flag and is separated
from the file by a space. Flags never contain "/" and a leading "/" is
used to determine whether the line has a flag or not.

Closes: #822462
Based-on-patch-by: Niels Thykier <niels@thykier.net>
Signed-off-by: Guillem Jover <guillem@debian.org>