libdpkg: Do not consider the database locked with a missing lock file
If the lock file does not exist, then in theory no other process is
supposed to have locked the database, as the lock file uses region
locking where the lock file is never supposed to be removed.

In practice, misguided users might end up removing the lock file in an
attempt to "unlock" the database (with potential catastrophic results)
but that is not a supported operation.

The supported scenario where the lock file is missing, is with a database
that has not yet been fully populated.

Reported-by: David Kalnischkies <donkult@debian.org>