test: Set PERL in the perl test suite
There is no guarantee that the PERL environment variable will contain
the perl executable path or name. When running the test suite from the
autotools build system we always set the PERL environment variable,
but on CPAN we were not doing that, and some CPAN testers have PERL
set to its version.