Dpkg::Source::Quilt: Add hint to check missing files on patch apply failures
When we are trying to patch a file that is not present in the original
tarball, the error message is rather confusing, as it mentions possible
fuzz problems, as if the file existed, which might mislead the user.

Add a hint to check whether the file is present in the original tarball
to avoid leading the user into dead ends.

Ideally the affected tarball would be reported in the error message, but
that information is currently not available on that code path.

Reported-by: Joseph Nahmias <jello@debian.org>