libdpkg: Use the amount of available memory instead phys_mem/2
On Linux there is the field 'MemAvailable' in '/proc/meminfo' which
holds the amount of memory which is available as of now. It considers
the fact that the page cache can be purged, some memory (if not all)
can be reclaimed (for instance by writing back dirty inodes) and some
memory should remain free just in case. This amount of memory can be
used without the need to swap-out some memory. The complete definition
can be located at [0].

The advantage over PHYS_MEM/2 is that it considers the current
status/usage of the system while assuming that half of what is
physically available is usable.

[guillem@debian.org:
 - Use fd_read() instead of read().
 - Merge two chained if blocks to reduce nesting level.
 - Add missing header.
 - Use a size_t const for the string length. ]

[0] https://www.kernel.org/doc/html/latest/filesystems/proc.html
Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>
Signed-off-by: Guillem Jover <guillem@debian.org>