dpkg: Fix --auto-deconfigure for essential and protected during installation
Only apply the --force-remove-essential and --force-remove-protected
during package removals, which can only happen due to Conflicts,
otherwise we cannot solve auto-deconfigurations due to Breaks on essential
or protected packages during installations or upgrades.

This is a regression when the Breaks field got introduced, as the same
function that had been used for Conflicts was refactored to be used for
Breaks, but without taking into account the removal case.

Fixes: commit b301c0e71a5314bb4560111c6cf1602269f6f672
Reported-by: Julian Andres Klode <jak@debian.org>
Ref: #983014