dpkg: Rework --assert-<feature> logic to be more robust
The --assert-<feature> options have traditionally required an installed
dpkg that has at least been configured once for the requested version,
as a way to signal that the dpkg that is running during the execution of
the check (usually within maintscripts, has a version that satisfies the
requirement).

This is a problem for installation bootstrapping, as there will be no
status file at that point.

Instead distinguish the two cases at hand, one for maintscripts, where
we check for the running dpkg version (the dpkg driving the current
operation), and the other when called in any other context, where the
support will be already in place and we can always return true.

Prompted-by: Helmut Grohne <helmut@subdivi.de>
Prompted-by: David Kalnischkies <donkult@debian.org>