dpkg-fsys-usrunmess: Move forced reconfiguration to the last step
Reconfiguring packages has the most potential for unaccounted breakage,
it is also the step that will take longer, and it should be the one
easier to recover from, by calling «dpkg --pending --configure» at a
later time. Move it at the end when we have done the reparation, and
performed the cleanup.

Ref: #991190
Stable-Candidate: 1.20.x