dpkg-fsys-usrunmess: Install a local policy-rc.d to ignore service restarts
When in systemd emergency mode, starting (including restarting) services
will force exiting the emergency mode, which can leave the system in a
pretty bad state. Part of the reason for reconfiguring all packages is
to both force any regeneration of missing files, and ideally to possibly
restart services so that they use the new real pathnames, but restarting
services is the most fragile part and with the most potential to fail
for whatever reason, so we will skip restarting services entirely. A
reboot after the conversion is strongly advised.

Closes: #991190
Stable-Candidate: 1.20.x