Dpkg::Shlibs::Symbol: Emit a warning on fully qualified symver patterns
When using the symver tag, we cannot use fully qualified patterns that
include a version tag and a symbol name, or the entry will not match.

This should ideally be an error, but that would produce build failures
for entries that might not currently be relevant for a build. Emit a
warning for now which can be promoted later on as an error.

Closes: #993991