build: Add a check for symlinks in the git repository
We generate dist tarballs with automake, which tells tar to dereference
symlinks. Check whether there is any symlink in the repository that will
degrade to a file and abort the tarball dist generation.

We ignore origins/default as even as a regular file, it works fine, and
it's less trouble to maintain for now.