dpkg: Restore fallback to "new-prerm failed-upgrade" for downgrades
This reverts commit 9d3ec0f5a727d439c3cbc8885bcbc78f7a34dd0b.

The change does not make much sense, as there are cases where the old
version does not need to have knowledge of the future problem to be able
to recover, when just not having the problematic code might be enough
for a correct installation.

This also removes this unexpected special-case keyed on the version that
is a completely anomalous compared with the rest of the maintscritpts
handling.

Analysis-by: Ian Jackson <ijackson@chiark.greenend.org.uk>
Closes: #996959