u-a: When initializing admindir from DPKG_ADMINDIR append "/alternatives"
When u-a is being run from within a maintscript, the DPKG_ADMINDIR
environment variable will contain the admindir for dpkg, which does not
include the "/alternatives" part, which we need to append ourselves,
otherwise we end up writing alternative state files into the base dpkg
database directory. This also meant that strange warnings being printed
about the state having changed, as no state was found in the wrong
directory.

Add a postinst maintscript to dpkg, to go over the dpkg database
directory entries, ignore the known ones, and move the ones that look
like an alternative state file back into the alternatives state
directory. This is going to be a short-lived fixup.

Closes: #1001198