build: Link all programs against libcompat
The libcompat library gets linked into libdpkg, and that works fine as
long as it is built only as a static library, because then any of the
libdpkg users in the project will pull in any required object from
there. But when libdpkg is built as a shared library, then libtool
does not even include it as a dependency in its .la tracking files.
And the visibility rules in the shared library will not expose any of
the libcompat symbols, so we need to link against it explicitly.

Prompted-by: Jörg Sonnenberger <joerg@NetBSD.org>