build: Rework TAP check hooking into the autotools machinery
Rename the TAP automake support file to tap.am to make it explicit this
is just one of the possible test suites we are running. Move from
implicitly hooking into the autotools check machinery to explicitly
hooking into it, so that we can run different test suites from the same
directory.