build: Add zlib-ng support
The zlib-ng library is a fork of zlib with its code modernized and lots
of optimizations for various modern CPUs giving it a substantial
performance boost.

This library provides either an API (and in theory ABI) compatible
interface, or a cleaned up API with namespaced and with better types.

We add support for the latter, via a compatibility header mapping symbol
names.