debian: Install optional localized man pages with dh_installman
We need to use a dh command so that debhelper can track which files got
installed.

We use the double globs, so that when a localized man page does not get
generated, we do not get failures from debhelper, as at least the
non-localized man page will always match.

We were doing this already some time ago, but dh_installman did not
handle localized man pages correctly and was overwriting them with the
last match. Now that this is fixed, we switch back to using the globs
within brace expansion to simplify the support and make use of a dh
command.

This reverts commit 806dfa8237d9e1e3c80a8b096cb1496d930fd34d.