u-a: Use the target timestamp when the symlink does not exist
Some programs track the symlink timestamps in their databases (such as
man-db). To be able to create reproducible installations we have to use
timestamps that are reproducible, at least on the first installation.

We also need to take into account that not using the current timestamp
when updating symlinks might cause backup systems to not work properly.

To cover these two cases, we will copy the timestamp from the target, if
the symlink did not exist.

Closes: #1004557