Dpkg::Source::Package: Print verifying signature for all source formats
Move the printing into the check_original_tarball_signature() function,
so that we can print one line per signature verified. Otherwise we could
only do that for the 1.0 format, which does not support multiple orig
tarballs.

While at it, simplify the message to only print the signature we are
verifying, and not the data file it is attached to.