debian: Warn in postinst about merged-/usr-via-aliased-dirs breakage
dpkg does not, and has never supported this filesystem layout even less
so the way it has been deployed as it bypasses dpkg entirely. It can
cause file loss, file overwrites, unexpected query results, among other
things, while being very insidious to detect if one is not aware. Adding
support for it would be rather infeasible and gets in the way of features
that have been on the works for some time now.