Dpkg::Index: Switch key function for control tests to be a stanza index
These stanzas have two possible distinguishing keys, the Tests and
Test-Command fields, but these are not necessarily unique, as can be
repeated given different constraints specified via other fields in the
stanza.

We switch to use the order tracking to index each stanza, so that we can
support duplicate test entries.

Reported-by: Paul Gevers <elbrus@debian.org>