build: Do not put localization entries on their own changelog group
The purpose of placing localization entries on their own changelog
groups was to try to reduce information overload, for something that
tends to be rather repetitive.

This is, though, a departure from the convention of using the person
taking responsibility for that change as that group header (within
square brackets). And not really needed given that we have switched
to standardized section names where we can place this following a more
uniform format.