debian: Clarify the merged-usr-via-aliased-dirs warning
It seems it was not obvious what the warning is about, even though the
FAQ entry seems rather clear (?). Mention that dpkg does not support it
(it has never), and the potential implications, so that the users are
aware of the problems, and not left in the dark of what they have been
set for. And can freely decide for themselves what to do, or not.

Prompted-by: Paul Wise <pabs@debian.org>