Dpkg::Compression: Conditionally use the gzip --rsyncable option
The gzip --rsyncable option is not universally supported, so we need to
conditionally use it. Ideally we would invoke «gzip --help» and check
whether the option is supported, but that would imply forking and
executing that process for any module that ends up loading this one,
which is not acceptable performance-wise. Instead we will approximate it
by Perl's osname, which is not ideal, but better than nothing.

Requires GNU gzip >= 1.7 for the --rsyncable option. On AIX GNU gzip
is too old. On the BSDs they use their own implementation based on zlib,
which does not currently support the --rsyncable option.

Fixes: commit 3a7db14e4a881d28fb73b29c126ea6f3ac4d0831
Fixes: https://rt.cpan.org/Ticket/Display.html?id=141805