test: Fix check output to assume zlib instead of zlib-ng
This is a quick workaround to get the builds back to nominal. We will
need to parametrize every size in outputs that depends on the compressor
used, as those are not under our control and could change between
different implementations (as happened here) or even same
implementations but different versions, if the compressors improve.

Fixes: commit 57d951260539d67da0230a83135ab805f89fa1af
Changelog: silent