Dpkg::Control::Fields: Add Static-Built-Using field support
This field mimics the previous Built-Using field semantics, but is
specifically intended for shadow dependencies stemming from static
builds.

In Debian, the Rust and Go teams agreed to use this language agnostic
field, instead of one for each of the languages. This means it can be
easily supported by dpkg, and can be used by other languages and
run-times.