man: Clarify uses cases for Static-Built-Using
Mention explicitly some of the use cases this field was thought for,
instead of letting it be the vague "static building", which can give
room to think it only applies to static linking. Where the intention
was for any type of static building, including also source-centered
language builds (like with Go or Rust), inclusion of header-only
libraries from C/C++, data blobs injection into code, etc.

Prompted-by: Shengjing Zhu <zhsj@debian.org>