test: Do not use non-POSIX shell syntax for arithmetic expressions
Use «$(( ))», instead of the non-POSIX syntax «(( ))», as in other parts
of the codebase. This fixes a test failure on Alpine Linux which uses
Busybox ash to execute the generated test script. Unfortunately this is
not being detected on systems with a dash POSIX shell, as it currently
disables its LINENO support so that GNU autotools configure does not end
up using it, and breaking due to bashisms (such as this one).

Based-on-patch-by: Sören Tempel <soeren@soeren-tempel.net>