debian: Silence warning about merged-usr-via-aliased-dirs on Debian
On Debian some people seem so offended by a (factual) warning, that the
obvious recourse has been for them to bully and abuse. Life's too short
for that.

If you might have wanted to be able to know about the status of dpkg
on *your* system, and decide for *yourself* what to do with it, sorry,
you will need to get aware some other way.