dpkg-statoverride: Add support for --force-not-root
When installing a stat override, we are changing the user/group and mode
for the pathname, which in most cases might require root privileges. Add
support for the not-root force option, so that it can be used standalone
or so that it does not break when called within a maintscript on a
not-root dpkg invocation.

Based-on-patch-by: John Spencer <maillist-dpkg@barfooze.de>
Closes: #1009069