Dpkg::Substvars: Add support for optional substvars assigned with ?=
Some packaging helpers (namely debhelper) set various substvars that
might usually be used by the packaging, but not necessarily always.

In the same way we mark some of the built-in variables as used, we add
a new operator «?=» to make it possible for other tools to mark those
other variables as optional too. Track those internally too so that they
can be printed back properly.

Proposed-by: Niels Thykier <niels@thykier.net>