dpkg-genbuildinfo: Add new can-execute-cross-built-programs tainted flag
Whether we can execute a cross-built program depends on whether the
CPU and the operating system have such capability, including in some
cases doing so via emulation, such as qemu hooked via something like
binfmt on Linux. If the system is or not capable of running such
programs, might affect the built artifacts and what tests run or not.

[guillem@debian.org:
 - Use File::Temp instead of tmpnam() and push_exit_handler().
 - Set a taint flag instead of a new field.
 - Pass -w to gcc to avoid warnings.
 - Refactor into a function.
 - Style fixes. ]

Closes: #1011191