Dpkg::Source::Package::V2: Always fix the permissions for upstream tarballs
The change to fix the directory traversal for source package unpacks
with in-place extractions of the debian.tar archive modified the way
that last extraction was done, to extract it also out-of-place, then
move the result into the destination. This had the consequence of
no longer fixing the permissions for the entire source tree, and instead
only for the debian/ directory. The previous calls for the orig tarballs
were not fixing up the permissions to avoid duplicating work, which
meant that now these did not get fixed any longer.

Remove the options that avoid fixing the permissions for all calls,
restoring the previous behavior.

Fixes: commit 7a6c03cb34d4a09f35df2f10779cbf1b70a5200b
Closes: #1012195
Stable-Candidates: 1.18.x 1.19.x 1.20.x
(cherry picked from commit 52d285fecbec8ba3cbd8255ef4de2be392d1e0de)