dpkg: Avoid setgid(2) call if not needed
If run as root, dpkg unconditionally calls setgid(0), for example if
called as «dpkg --print-foreign-architectures» via lsb_release(1). On
Linux this causes a POSIX capability check for CAP_SETGID, which for
SELinux performs a permission check on the subject type. Allowing that
access unnecessarily broadens the privileges of the subject, and ignoring
(dontaudit'ing) might hide situations where CAP_SETGID is required
(maybe within another program, but same SELinux type).

This change makes the code only call setgid() if the current group ID
is not 0 (=root). Since the condition also checks for the root user
via getuid(), checking getegid() seems not to be necessary.

Closes: #1014332
Signed-off-by: Guillem Jover <guillem@debian.org>