Dpkg::OpenPGP: Add Stateless OpenPGP (SOP) backend support
This adds support for the Stateless OpenPGP Command-Line Interface,
as drafted at:

  https://datatracker.ietf.org/doc/draft-dkg-openpgp-stateless-cli/

as a new OpenPGP backend. Multiple SOP implementations can be supported,
as long as they support the required draft version and interfaces.

This commit currently only explicitly probes for the pgpainless-cli
SOP implementation, as the one that conforms to our expectations.