dpkg-buildpackage: Add --sign-backend to specify the OpenPGP backend
The default OpenPGP backend is auto-detected, but can now be specified
explicitly.

There was a historical -s(gpg|pgp) option which we are not reusing as it
was confusing with the options that control the source package contents.
This should be rare enough to be needed on the command-line that a long
option should be fine, and it can always be configured on the config
file.