man: Use stanza instead of paragraph, block or record for deb822 parts
The «stanza» name is a commonly used and understood term when referring
to deb822 parts. Although «paragraph» is commonly used it has the
problem of being confusing as it then makes it hard to distinguish
actual text paragraphs in prose, while «stanza» is a very specific
term that is not applied anywhere else in the deb822 context, so
it's always more clear and specific, than also block or record.

In addition «stanza» is shorter, which is always a nice attribute
on code for example.

This unifies the terminology used with recent changes in the Debian
policy manual, to try to make these concepts more clear in general.