man: Clarify when package names get arch-qualified in dpkg-query output
There were two places describing when packages get arch-qualified,
one was missing the foreign arch case, the other read as if it was a
non-exhaustive list of cases when it already was, which was rather
confusing.

Use the opportunity the reword both in a similar and hopefully more
clear way and expand what a foreign arch implies.

Reported-by: Jakub Wilk <jwilk@jwilk.net> (on IRC)