Dpkg::Vendor: Fix get_vendor_file() to return on first match
There is no point in checking the whole list of alternative vendor
files just to return the last match, when we can simply return early
on the first, as that's all wasted work. This also makes the behavior
consistent with the get_vendor_object which returns on the first match
when looking for a perl module.

Reported-by: Niels Thykier <niels@thykier.net> (on IRC)