Dpkg::Vendor: Fix vendor file loading for vendors with special characters
The current code was not handling vendor names with multiple groups
of spaces nor with special characters (any of [^A-Za-z0-9]). While
filenames can be easily supported by many of those special characters,
they can be annoying to deal with or cause unexpected results (such as
using «/»). We add trials with full special character remapping to «-»
using the same regex that we will use for vendor module names, so that
the remapping is unified and more easy to understand.