dselect: Require the Dpkg modules for the ftp access method
While strictly speaking we do not currently need it, it will make it
possible to remove code that can use functionality from the Dpkg perl
modules. We replace the warning recommending to install perl, by a
message stating that the Dpkg modules are missing, as if the user
installs libdpkg-perl (or equivalent) then the perl package will also
be pulled in.