dselect: Unconditionally load perl modules included in perl-base
In Debian and derivatives where the perl distribution is split into
a perl-base package which is always guaranteed to be installed, and a
perl-modules and perl packages, which are not, we do not need to
conditionally load these modules, as they will always be installed.

On other systems, when perl is installed these modules will always be
present so this change should not make any difference.

Changelog: internal