Dselect::Ftp: Conditionally load perl modules shipped in the perl package
In Debian and derivatives where the perl distribution is split into
a perl-base package which is always guaranteed to be installed, and a
perl-modules and perl packages, which are not, we need to conditionally
load these modules, as we do not have strict dependencies on the perl
package.

On other systems, when perl is installed these modules will always be
present so this change should not make any difference.