Dpkg::OpenPGP: Change inline_verify to take an output file
When doing inline-verify, ideally we should let the OpenPGP
implementation verify the inlinesigned data and output the verified data
with the ASCII Armor and any heading or trailing junk stripped.

We are currently not doing that, as we cannot assume we might have an
OpenPGP implementation around when we need to parse deb822 data, so for
now we will just pass undef as the output filename, but in the future
we might rearchitect the code to make use of this.

At least this way the implementation can be used properly by potential
future users, once and if the module becomes public, instead of
providing an insecure API.

Requested-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>