libdpkg: Fix available memory computation for xz thread scaling
On Linux, the memory will tend to get used up completely by caches and
buffers for performance reasons. The Available memory thus tends to be
close to zero. Basing the thread number on that measure of available
memory is as such not very useful, and has caused build systems to fall
back to serial compression, making it extremely slow, in same cases
reported to take hours.

Gather the MemFree, Buffers and Cached fields from the /proc/meminfo
file and add them up as the computed available memory which is what is
really available to be used by the system when needed.

Closes: #1023870