u-a: Switch point at which we increment alternative refcount
Instead of incrementing the refcount from the call sites, do that from
within the alternative_map_add() function which is the one taking
the reference. And unref on the functions that get their local functions
get out of scope. Which is way more correct, and removes the anti-pattern
of having to recall to increment the refcount before calling any
subsequent alternative_map_add() after the second one.