libdpkg: Automatically initialize fsys and db dirs on demand
Change the dpkg_fsys_* and dpkg_db_* functions to initialize the
directories when needed, so that it guarantees the users can always
use these functions safely, and they respect the expected initialization
order.

Refactor the set_instdir() and set_admindir() functions so that we can
set their values correctly in a single place.

Fixes: commit b8f9e1c3ce86818d14ccad403e6ad2e6b52e6b12
Reported-by: stackomatiker (on stackoverflow)
Ref: https://stackoverflow.com/questions/73529576