dpkg-architecture: Properly compute required information to gather
We are using bits to track the required information to gather. The
problem was that we were also using those same bits to discriminate
what variables had been seen from the environment and to remove them
if so, but those bits were not unique per variable, several were shared
among various variables, causing information to not be gathered.

Instead track seen variables with a hash, and move the dependencies
back into the environment variable definitions. We also use
high-precedence operators on the conditional for the next command.

This was surfaced when using GNU make >= 4.4, as that respects exported
make variables into $(shell) functions, which caused infinite loops due
to various of the calls in the test suite not returning anything anymore.

Reported-by: Sören Tempel <soeren@soeren-tempel.net>