test: Use perl instead of yes+dd to generate a data-file for deb-split
Using yes+dd is problematic, and we already had to cope with SIGPIPE
issues in the past. When using busybox's dd, the output is truncated
unless we use iflag=fullblock which is a rather non-portable option.

POSIX in general is rather lacking when it comes to standard commands
to fetch a limited number of bytes. Instead of trying to do more
portability gymnastics, simply switch to use perl, which we already
depend on for the test suite, and is going to be more portable and
featureful.