libdpkg: Fix pdb_lax_stanza_parser and pdb_allow_empty aliased values
The pdb_lax_stanza_parser constant was assigned the last value to avoid
renumbering all other enum constants, but because it was placed
unordered it was predictably not future-proof. Later on a new constant
was assigned as the last enum constant reusing that same value, causing
the current aliasing.

This ends up currently being a non-issue, as all places where we pass
one of the constants we also pass the other.

Fixes: commit 99c9cdbf1b9650ef5025c3c06ea918ef3007888e