debian: Stop shipping the «git log» changelog in exchange for no trimming
It does not make much sense to ship the upstream «git log» changelog
which contains lots of fluff and metadata. At the same time the recent
debhelper trimming is not ideal as there's no easy way to fetch old
entries anymore. So stop shipping the upstream changelog in favor of
not trimming the debian/changelog one, which acts as a proper user
readable changelog.

Closes: #1027716